+++ lib/rubocop/cop/style/empty_lines_around_access_modifier.rb
@@ -14,15 +14,26 @@ module Rubocop
         def on_send(node)
           return unless modifier_node?(node)
 
-          send_line = node.loc.line
+          return if empty_lines_around?(node)
 
-          unless processed_source[send_line].blank? &&
-              processed_source[send_line - 2].blank?
-            convention(node, :expression)
-          end
+          convention(node, :expression)
         end
 
         private
+
+        def empty_lines_around?(node)
+          send_line = node.loc.line
+          previous_line = processed_source[send_line - 2]
+          next_line = processed_source[send_line]
+
+          (class_def?(previous_line.lstrip) ||
+           previous_line.blank?) &&
+            next_line.blank?
+        end
+
+        def class_def?(line)
+          %w(class module).any? { |keyword| line.start_with?(keyword) }
+        end
 
         def message(node)
           format(MSG, node.loc.selector.source)